requirements.txt
tiktoken
numpy
faiss-cpu
packaging